final fixes :)

+++ src/ui/settings/nvsSettings.cpp
@@ -4,8 +4,8 @@ bool RTC_DATA_ATTR disableAllVibration = false;
 bool RTC_DATA_ATTR disableWakeUp = false;
 
 void loadAllStorage() {
-    disableAllVibration = bool(fsGetString(CONF_DISABLE_ALL_VIBRATION, "0"));
-    disableWakeUp = bool(fsGetString(CONF_DISABLE_WAKE_UP, "0"));
+    disableAllVibration = bool(fsGetString(CONF_DISABLE_ALL_VIBRATION, "0").toInt());
+    disableWakeUp = bool(fsGetString(CONF_DISABLE_WAKE_UP, "0").toInt());
 
     debugLog("disableAllVibration: " + BOOL_STR(disableAllVibration));
     debugLog("disableWakeUp: " + BOOL_STR(disableWakeUp));
